Reorganize DbWriter to multiple classes for clarity
