+++ sink-connector/src/main/java/com/altinity/clickhouse/sink/connector/db/operations/ClickHouseAlterTable.java
@@ -1,7 +1,13 @@
 package com.altinity.clickhouse.sink.connector.db.operations;
 
 import com.altinity.clickhouse.sink.connector.db.ClickHouseDbConstants;
+import com.clickhouse.jdbc.ClickHouseConnection;
+import org.apache.kafka.connect.data.Field;
+import org.slf4j.Logger;
+import org.slf4j.LoggerFactory;
 
+import java.util.ArrayList;
+import java.util.List;
 import java.util.Map;
 
 /**
@@ -9,6 +15,7 @@ import java.util.Map;
  * in ClickHouse(Schema Evolution)
  */
 public class ClickHouseAlterTable extends ClickHouseTableOperationsBase{
+    private static final Logger log = LoggerFactory.getLogger(ClickHouseAlterTable.class.getName());
 
     public enum ALTER_TABLE_OPERATION {
         ADD("add"),
@@ -38,7 +45,42 @@ public class ClickHouseAlterTable extends ClickHouseTableOperationsBase{
         }
         alterTableSyntax.deleteCharAt(alterTableSyntax.lastIndexOf(","));
 
-
         return alterTableSyntax.toString();
+    }
+
+    /**
+     *
+     * @para
+     * m modifiedFields
+     */
+    public void alterTable(List<Field> modifiedFields, String tableName, ClickHouseConnection connection, Map<String, String> columnNameToDataTypeMap) {
+        List<Field> missingFieldsInCH = new ArrayList<Field>();
+        // Identify the columns that need to be added/removed in ClickHouse.
+        for(Field f: modifiedFields) {
+            String colName = f.name();
+
+            if(columnNameToDataTypeMap.containsKey(colName) == false) {
+                missingFieldsInCH.add(f);
+            }
+        }
+
+        if(!missingFieldsInCH.isEmpty()) {
+            log.info("***** ALTER TABLE ****");
+            ClickHouseAlterTable cat = new ClickHouseAlterTable();
+            Field[] missingFieldsArray = new Field[missingFieldsInCH.size()];
+            missingFieldsInCH.toArray(missingFieldsArray);
+            Map<String, String> colNameToDataTypeMap = cat.getColumnNameToCHDataTypeMapping(missingFieldsArray);
+
+            if(!colNameToDataTypeMap.isEmpty()) {
+                String alterTableQuery = cat.createAlterTableSyntax(tableName, colNameToDataTypeMap, ClickHouseAlterTable.ALTER_TABLE_OPERATION.ADD);
+                log.info(" ***** ALTER TABLE QUERY **** " + alterTableQuery);
+
+                try {
+                    cat.runQuery(alterTableQuery, connection);
+                } catch(Exception e) {
+                    log.error(" **** ALTER TABLE EXCEPTION ", e);
+                }
+            }
+        }
     }
 }
